Update: extract Header style
