+++ pages/index.js
@@ -1,4 +1,4 @@
-import Header from 'components/Header'
+import Header from 'components/Header/Header'
 
 export default function Home() {
   return <Header />
